requirements.txt
flask
google-genai
validators
